refactor(status): extract lightweight token formatter
